+++ components/FontSizeSelector.tsx
@@ -7,11 +7,12 @@ const maxValue = 50
 
 interface Props {
   fontSize: number,
+  disabled: boolean,
   onChange: (fontSize: number) => void,
   onHover: (inside: boolean) => void,
 }
 
-export default function FontSizeSelector({fontSize, onChange, onHover}: Props) {
+export default function FontSizeSelector({fontSize, disabled, onChange, onHover}: Props) {
   const [showTooltip, setShowTooltip] = useState(false)
   return (
     <Slider
@@ -21,6 +22,7 @@ export default function FontSizeSelector({fontSize, onChange, onHover}: Props) {
       max={maxValue}
       colorScheme='blue'
       value={fontSize}
+      isDisabled={disabled}
       onChange={(v) => onChange(v)}
       onMouseEnter={() => {
         setShowTooltip(true)