+++ model/ReaderSettings.ts
@@ -3,6 +3,7 @@ import {ReadingDirection} from "./ReadingDirection"
 
 export interface ReaderSettings {
   zoom: number,
+  autoFontSize: boolean,
   fontSize: number,
   showText: boolean,
   showParagraphs: boolean,